KA-269 Rename reducers to .ts
